# Conflicts:
#	src/1. Layers/1.3 Domain/Well.Domain/ValueObjects/DeliveryLine.cs

+++ src/1. Layers/1.3 Domain/Well.Domain/ValueObjects/DeliveryLine.cs
@@ -20,8 +20,7 @@
         public string ProductCode { get; set; }
 
         public string ProductDescription { get; set; }
-
-        public int Value { get; set; }
+        public decimal Value { get; set; }
 
         public int InvoicedQuantity { get; set; }
